+++ src/snowflake/cli/_app/snow_connector.py
@@ -39,7 +39,6 @@ from snowflake.cli.api.config import (
 from snowflake.cli.api.constants import DEFAULT_SIZE_LIMIT_MB
 from snowflake.cli.api.exceptions import (
     CliError,
-    InvalidConnectionConfigurationError,
     SnowflakeConnectionError,
 )
 from snowflake.cli.api.feature_flags import FeatureFlag
@@ -48,7 +47,7 @@ from snowflake.cli.api.secure_path import SecurePath
 from snowflake.cli.api.utils.types import try_cast_to_bool
 from snowflake.connector import SnowflakeConnection
 from snowflake.connector.auth.workload_identity import ApiFederatedAuthenticationType
-from snowflake.connector.errors import DatabaseError
+from snowflake.connector.errors import DatabaseError, ForbiddenError
 
 log = logging.getLogger(__name__)
 
@@ -190,8 +189,6 @@ def connect_to_snowflake(
     connection_name: Optional[str] = None,
     **overrides,
 ) -> SnowflakeConnection:
-    from snowflake.connector.errors import ForbiddenError
-
     if temporary_connection and connection_name:
         raise ClickException("Can't use connection name and temporary connection.")
     elif not temporary_connection and not connection_name:
@@ -305,10 +302,8 @@ def connect_to_snowflake(
                 application=command_info(),
                 **connection_parameters,
             )
-    except ForbiddenError as err:
-        raise SnowflakeConnectionError(err)
-    except DatabaseError as err:
-        raise InvalidConnectionConfigurationError(err.msg)
+    except (ForbiddenError, DatabaseError) as err:
+        raise SnowflakeConnectionError(err) from err
 
 
 def _avoid_closing_the_connection_if_it_was_shared(